Nexus: Rescale the scene on wallpaper resize
When device is rotated and screen orientation changes from portrait
to landscape (or vice versa), the background image is rescaled to
the new size without taking into account screen aspect ratio. This
results in squeezed or stretched pyramids, and in some cases the
image may no longer cover the entire workspace.

This patch adds to the scene a world scale that is re-calculated
on resize taking into account the new screen proportions.

Change-Id: Id98a49fc4506dfe810d5bfa224b609d6940f7c05
Signed-off-by: Mykola Ostrovskyy <mykola@ti.com>